# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_handleRefreshToken","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/auth/account":{"get":{"operationId":"AuthController_handleGetAccount","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_handleLogout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/roles":{"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["roles"]},"get":{"operationId":"RolesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/v1/roles/{id}":{"get":{"operationId":"RolesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["roles"]},"put":{"operationId":"RolesController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["roles"]},"delete":{"operationId":"RolesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/v1/companies":{"post":{"operationId":"CompaniesController_create","summary":"Admin create new user","description":"\n* Only admin can use this API\n\n* Admin create user and give some specific information","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"},"examples":{"user_1":{"value":{"first_name":"John","last_name":"Doe","email":"johndoe@example.com","password":"1232@asdS"}},"user_2":{"value":{"first_name":"Michael","last_name":"Smith","email":"michaelsmith@example.com","password":"1232@asdS"}}}}}},"responses":{"201":{"description":""}},"tags":["companies"]},"get":{"operationId":"CompaniesController_findAll","parameters":[{"name":"pageSize","required":false,"in":"query","examples":{"10":{"value":10,"description":"Get 10 collection"},"50":{"value":50,"description":"Get 50 collection"}},"schema":{"type":"number"}},{"name":"current","required":true,"in":"query","examples":{"1":{"value":1,"description":"Start from page 1"},"10":{"value":10,"description":"Start from page 10"}},"schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["companies"]}},"/api/v1/companies/{id}":{"get":{"operationId":"CompaniesController_findOne","parameters":[{"name":"id","required":true,"in":"path","examples":{"migration_id_1":{"value":"644293b09150e9f67d9bb75d","description":"Collection Kitchen vocabulary"},"migration_id_2":{"value":"6442941027467f9a755ff76d","description":"Collection Sport vocabulary"}},"schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["companies"]},"put":{"operationId":"CompaniesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":""}},"tags":["companies"]},"delete":{"operationId":"CompaniesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["companies"]}},"/api/v1/jobs":{"post":{"operationId":"JobsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateJobDto"}}}},"responses":{"201":{"description":""}},"tags":["jobs"]},"get":{"operationId":"JobsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["jobs"]}},"/api/v1/jobs/{id}":{"get":{"operationId":"JobsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["jobs"]},"put":{"operationId":"JobsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateJobDto"}}}},"responses":{"200":{"description":""}},"tags":["jobs"]},"delete":{"operationId":"JobsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["jobs"]}},"/api/v1/files/upload":{"post":{"operationId":"FilesController_uploadFile","summary":"User create their collection","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["files"],"security":[{"JWT-auth":[]}]}},"/api/v1/files":{"post":{"operationId":"FilesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFileDto"}}}},"responses":{"201":{"description":""}},"tags":["files"]},"get":{"operationId":"FilesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["files"]}},"/api/v1/files/{id}":{"get":{"operationId":"FilesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["files"]},"patch":{"operationId":"FilesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFileDto"}}}},"responses":{"200":{"description":""}},"tags":["files"]},"delete":{"operationId":"FilesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["files"]}},"/api/v1/permissions":{"post":{"operationId":"PermissionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePermissionDto"}}}},"responses":{"201":{"description":""}},"tags":["permissions"]},"get":{"operationId":"PermissionsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["permissions"]}},"/api/v1/permissions/by-admin":{"get":{"operationId":"PermissionsController_findAllByAdmin","parameters":[],"responses":{"200":{"description":""}},"tags":["permissions"]}},"/api/v1/permissions/{id}":{"get":{"operationId":"PermissionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["permissions"]},"put":{"operationId":"PermissionsController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePermissionDto"}}}},"responses":{"200":{"description":""}},"tags":["permissions"]},"delete":{"operationId":"PermissionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["permissions"]}},"/api/v1/resumes":{"post":{"operationId":"ResumesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserCvDto"}}}},"responses":{"201":{"description":""}},"tags":["resumes"]},"get":{"operationId":"ResumesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["resumes"]}},"/api/v1/resumes/by-user":{"post":{"operationId":"ResumesController_getResumesByUser","parameters":[],"responses":{"201":{"description":""}},"tags":["resumes"]}},"/api/v1/resumes/{id}":{"get":{"operationId":"ResumesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["resumes"]},"put":{"operationId":"ResumesController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResumeDto"}}}},"responses":{"200":{"description":""}},"tags":["resumes"]},"delete":{"operationId":"ResumesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["resumes"]}},"/api/v1/subscribers":{"get":{"operationId":"SubscribersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["subscribers"]},"post":{"operationId":"SubscribersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubscriberDto"}}}},"responses":{"201":{"description":""}},"tags":["subscribers"]},"put":{"operationId":"SubscribersController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubscriberDto"}}}},"responses":{"200":{"description":""}},"tags":["subscribers"]}},"/api/v1/subscribers/skills":{"post":{"operationId":"SubscribersController_getUserSkills","parameters":[],"responses":{"201":{"description":""}},"tags":["subscribers"]}},"/api/v1/subscribers/{id}":{"get":{"operationId":"SubscribersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["subscribers"]},"delete":{"operationId":"SubscribersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["subscribers"]}},"/api/v1/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}}}},"/api/v1/mail/send-job-notification":{"post":{"operationId":"MailController_sendJobNotification","parameters":[],"responses":{"201":{"description":""}},"tags":["mail"]}},"/api/v1/mail":{"get":{"operationId":"MailController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["mail"]}},"/api/v1/mail/{id}":{"get":{"operationId":"MailController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["mail"]},"patch":{"operationId":"MailController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMailDto"}}}},"responses":{"200":{"description":""}},"tags":["mail"]},"delete":{"operationId":"MailController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["mail"]}}},"info":{"title":"Nest JS Advanced Course for developers [2024]","description":"## API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"type":"http","scheme":"bearer"}},"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"RegisterUserDto":{"type":"object","properties":{}},"CreateRoleDto":{"type":"object","properties":{}},"UpdateRoleDto":{"type":"object","properties":{}},"CreateCompanyDto":{"type":"object","properties":{}},"UpdateCompanyDto":{"type":"object","properties":{}},"CreateJobDto":{"type":"object","properties":{}},"UpdateJobDto":{"type":"object","properties":{}},"CreateFileDto":{"type":"object","properties":{}},"UpdateFileDto":{"type":"object","properties":{}},"CreatePermissionDto":{"type":"object","properties":{}},"UpdatePermissionDto":{"type":"object","properties":{}},"CreateUserCvDto":{"type":"object","properties":{}},"UpdateResumeDto":{"type":"object","properties":{}},"CreateSubscriberDto":{"type":"object","properties":{}},"UpdateSubscriberDto":{"type":"object","properties":{}},"UpdateMailDto":{"type":"object","properties":{}}}}}